+++ packages/manager/src/components/AccountActivation/accountActivationLandingLazyRoute.ts
@@ -1,0 +1,9 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { AccountActivationLanding } from 'src/components/AccountActivation/AccountActivationLanding';
+
+export const accountActivationLandingLazyRoute = createLazyRoute(
+  '/account-activation'
+)({
+  component: AccountActivationLanding,
+});
