+++ packages/manager/src/features/Domains/CreateDomain/createDomainLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { CreateDomain } from 'src/features/Domains/CreateDomain/CreateDomain';
+
+export const createDomainLazyRoute = createLazyRoute('/domains/create')({
+  component: CreateDomain,
+});
