+++ packages/manager/src/features/Domains/DomainDetail/domainDetailLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { DomainDetail } from 'src/features/Domains/DomainDetail/DomainDetail';
+
+export const domainDetailLazyRoute = createLazyRoute('/domains/$domainId')({
+  component: DomainDetail,
+});
