+++ packages/manager/src/features/Domains/domainsLandingLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { DomainsLanding } from 'src/features/Domains/DomainsLanding';
+
+export const domainsLandingLazyRoute = createLazyRoute('/domains')({
+  component: DomainsLanding,
+});
