+++ packages/manager/src/features/Events/eventsLandingLazyRoute.ts
