+++ packages/manager/src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { FirewallDetail } from 'src/features/Firewalls/FirewallDetail';
+
+export const firewallDetailLazyRoute = createLazyRoute('/firewalls/$id')({
+  component: FirewallDetail,
+});
