+++ packages/manager/src/features/IAM/Users/userDetailsLandingLazyRoute.ts
@@ -1,11 +1,6 @@
 import { createLazyRoute } from '@tanstack/react-router';
 
-import { IdentityAccessLanding } from 'src/features/IAM/IAMLanding';
 import { UserDetailsLanding } from 'src/features/IAM/Users/UserDetailsLanding';
-
-export const iamLandingLazyRoute = createLazyRoute('/iam')({
-  component: IdentityAccessLanding,
-});
 
 export const userDetailsLandingLazyRoute = createLazyRoute(
   '/iam/users/$username'
