+++ packages/manager/src/features/IAM/iamLandingLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { IdentityAccessLanding } from 'src/features/IAM/IAMLanding';
+
+export const iamLandingLazyRoute = createLazyRoute('/iam')({
+  component: IdentityAccessLanding,
+});
