+++ packages/manager/src/features/Images/ImagesLanding/imagesLandingLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { ImagesLanding } from 'src/features/Images/ImagesLanding/ImagesLanding';
+
+export const imagesLandingLazyRoute = createLazyRoute('/images')({
+  component: ImagesLanding,
+});
