+++ packages/manager/src/features/Kubernetes/KubernetesClusterDetail/kubernetesClusterDetailLazyRoute.ts
@@ -1,0 +1,9 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { KubernetesClusterDetail } from 'src/features/Kubernetes/KubernetesClusterDetail/KubernetesClusterDetail';
+
+export const kubernetesClusterDetailLazyRoute = createLazyRoute(
+  '/kubernetes/clusters/$clusterId'
+)({
+  component: KubernetesClusterDetail,
+});
