+++ packages/manager/src/features/Longview/LongviewDetail/LongviewDetailWrapper.tsx
@@ -1,17 +1,12 @@
-import { createLazyRoute, useParams } from '@tanstack/react-router';
+import { useParams } from '@tanstack/react-router';
 import * as React from 'react';
 
 import EnhancedLongviewDetail from 'src/features/Longview/LongviewDetail/LongviewDetail';
-import LongviewLanding from 'src/features/Longview/LongviewLanding/LongviewLanding';
-
-export const longviewLandingLazyRoute = createLazyRoute('/longview')({
-  component: LongviewLanding,
-});
 
 // Making a functional component to wrap the EnhancedLongviewDetail HOC
 // Ideally we would refactor this and fetch the data properly but considering Longview is nearing its end of life
 // we'll just match the legacy routing behavior
-const LongviewDetailWrapper = () => {
+export const LongviewDetailWrapper = () => {
   const { id } = useParams({ from: '/longview/clients/$id' });
 
   if (!id) {
@@ -33,9 +28,3 @@ const LongviewDetailWrapper = () => {
 
   return <EnhancedLongviewDetail {...props} />;
 };
-
-export const longviewDetailLazyRoute = createLazyRoute('/longview/clients/$id')(
-  {
-    component: LongviewDetailWrapper,
-  }
-);
