+++ packages/manager/src/features/Longview/LongviewLanding/longviewLandingLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import LongviewLanding from 'src/features/Longview/LongviewLanding/LongviewLanding';
+
+export const longviewLandingLazyRoute = createLazyRoute('/longview')({
+  component: LongviewLanding,
+});
