+++ packages/manager/src/features/ObjectStorage/objectStorageLandingLazyRoute.ts
@@ -1,6 +1,5 @@
 import { createLazyRoute } from '@tanstack/react-router';
 
-import { BucketDetailLanding } from 'src/features/ObjectStorage/BucketDetail';
 import { ObjectStorageLanding } from 'src/features/ObjectStorage/ObjectStorageLanding';
 
 export const objectStorageLandingLazyRoute = createLazyRoute('/object-storage')(
@@ -8,9 +7,3 @@ export const objectStorageLandingLazyRoute = createLazyRoute('/object-storage')(
     component: ObjectStorageLanding,
   }
 );
-
-export const bucketDetailLandingLazyRoute = createLazyRoute(
-  '/object-storage/buckets/$clusterId/$bucketName'
-)({
-  component: BucketDetailLanding,
-});
