+++ packages/manager/src/features/PlacementGroups/PlacementGroupsDetail/placementGroupDetailLazyRoute.ts
@@ -1,0 +1,9 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { PlacementGroupsDetail } from 'src/features/PlacementGroups/PlacementGroupsDetail/PlacementGroupsDetail';
+
+export const placementGroupsDetailLazyRoute = createLazyRoute(
+  '/placement-groups/$id'
+)({
+  component: PlacementGroupsDetail,
+});
