+++ packages/manager/src/features/Support/SupportTicketDetail/supportTicketDetailLazyRoute.ts
@@ -1,0 +1,9 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { SupportTicketDetail } from './SupportTicketDetail';
+
+export const supportTicketDetailLazyRoute = createLazyRoute(
+  '/support/tickets/$ticketId'
+)({
+  component: SupportTicketDetail,
+});
