+++ packages/manager/src/features/Support/SupportTickets/supportTicketsLandingLazyRoute.ts
@@ -1,0 +1,9 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { SupportTicketsLanding } from 'src/features/Support/SupportTickets/SupportTicketsLanding';
+
+export const supportTicketsLandingLazyRoute = createLazyRoute(
+  '/support/tickets'
+)({
+  component: SupportTicketsLanding,
+});
