+++ packages/manager/src/features/VPCs/VPCCreate/VPCCreateLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import VPCCreate from 'src/features/VPCs/VPCCreate/VPCCreate';
+
+export const vpcCreateLazyRoute = createLazyRoute('/vpcs/create')({
+  component: VPCCreate,
+});
