+++ packages/manager/src/features/VPCs/VPCLanding/VPCLandingLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import VPCLanding from 'src/features/VPCs/VPCLanding/VPCLanding';
+
+export const vpcLandingLazyRoute = createLazyRoute('/')({
+  component: VPCLanding,
+});
