+++ packages/manager/src/features/Volumes/volumesCreateLazyRoute.ts
@@ -1,11 +1,6 @@
 import { createLazyRoute } from '@tanstack/react-router';
 
 import { VolumeCreate } from 'src/features/Volumes/VolumeCreate';
-import { VolumesLanding } from 'src/features/Volumes/VolumesLanding';
-
-export const volumesLandingLazyRoute = createLazyRoute('/')({
-  component: VolumesLanding,
-});
 
 export const volumeCreateLazyRoute = createLazyRoute('/volumes/create')({
   component: VolumeCreate,
