+++ packages/manager/src/features/Volumes/volumesLandingLazyRoute.ts
@@ -1,0 +1,7 @@
+import { createLazyRoute } from '@tanstack/react-router';
+
+import { VolumesLanding } from 'src/features/Volumes/VolumesLanding';
+
+export const volumesLandingLazyRoute = createLazyRoute('/')({
+  component: VolumesLanding,
+});
