+++ packages/manager/src/routes/firewalls/index.ts
@@ -20,7 +20,9 @@ const firewallsIndexRoute = createRoute({
   getParentRoute: () => firewallsRoute,
   path: '/',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallLandingLazyRoute)
+  import(
+    'src/features/Firewalls/FirewallLanding/firewallLandingLazyRoute'
+  ).then((m) => m.firewallLandingLazyRoute)
 );
 
 const firewallCreateRoute = createRoute({
@@ -28,7 +30,9 @@ const firewallCreateRoute = createRoute({
   path: 'create',
   validateSearch: (search: FirewallsSearchParams) => search,
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallLandingLazyRoute)
+  import(
+    'src/features/Firewalls/FirewallLanding/firewallLandingLazyRoute'
+  ).then((m) => m.firewallLandingLazyRoute)
 );
 
 const firewallDetailRoute = createRoute({
@@ -50,91 +54,117 @@ const firewallDetailRoute = createRoute({
   path: '$id',
   validateSearch: (search: { tab?: string }) => search,
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailRulesRoute = createRoute({
   getParentRoute: () => firewallsRoute,
   path: '$id/rules',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailRulesAddRuleRoute = createRoute({
   getParentRoute: () => firewallDetailRulesRoute,
   path: 'add',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailRulesEditInboundRuleRoute = createRoute({
   getParentRoute: () => firewallDetailRulesRoute,
   path: 'edit/inbound/$ruleId',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailRulesEditOutboundRuleRoute = createRoute({
   getParentRoute: () => firewallDetailRulesRoute,
   path: 'edit/outbound/$ruleId',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailRulesAddInboundRuleRoute = createRoute({
   getParentRoute: () => firewallDetailRulesAddRuleRoute,
   path: 'inbound',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailRulesAddOutboundRuleRoute = createRoute({
   getParentRoute: () => firewallDetailRulesAddRuleRoute,
   path: 'outbound',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailLinodesRoute = createRoute({
   getParentRoute: () => firewallsRoute,
   path: '$id/linodes',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailLinodesAddLinodeRoute = createRoute({
   getParentRoute: () => firewallDetailLinodesRoute,
   path: 'add',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailLinodesRemoveLinodeRoute = createRoute({
   getParentRoute: () => firewallDetailLinodesRoute,
   path: 'remove',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailNodebalancersRoute = createRoute({
   getParentRoute: () => firewallsRoute,
   path: '$id/nodebalancers',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailNodebalancersAddNodebalancerRoute = createRoute({
   getParentRoute: () => firewallDetailNodebalancersRoute,
   path: 'add',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 const firewallDetailNodebalancersRemoveNodebalancerRoute = createRoute({
   getParentRoute: () => firewallDetailNodebalancersRoute,
   path: 'remove',
 }).lazy(() =>
-  import('./firewallLazyRoutes').then((m) => m.firewallDetailLazyRoute)
+  import('src/features/Firewalls/FirewallDetail/firewallDetailLazyRoute').then(
+    (m) => m.firewallDetailLazyRoute
+  )
 );
 
 export const firewallsRouteTree = firewallsRoute.addChildren([
