+++ packages/manager/src/routes/objectStorage/index.ts
@@ -20,7 +20,7 @@ const objectStorageIndexRoute = createRoute({
   getParentRoute: () => objectStorageRoute,
   path: '/',
 }).lazy(() =>
-  import('./objectStorageLazyRoutes').then(
+  import('src/features/ObjectStorage/objectStorageLandingLazyRoute').then(
     (m) => m.objectStorageLandingLazyRoute
   )
 );
@@ -29,7 +29,7 @@ const objectStorageBucketsLandingRoute = createRoute({
   getParentRoute: () => objectStorageRoute,
   path: 'buckets',
 }).lazy(() =>
-  import('./objectStorageLazyRoutes').then(
+  import('src/features/ObjectStorage/objectStorageLandingLazyRoute').then(
     (m) => m.objectStorageLandingLazyRoute
   )
 );
@@ -38,7 +38,7 @@ const objectStorageAccessKeysLandingRoute = createRoute({
   getParentRoute: () => objectStorageRoute,
   path: 'access-keys',
 }).lazy(() =>
-  import('./objectStorageLazyRoutes').then(
+  import('src/features/ObjectStorage/objectStorageLandingLazyRoute').then(
     (m) => m.objectStorageLandingLazyRoute
   )
 );
@@ -47,7 +47,7 @@ const objectStorageBucketCreateRoute = createRoute({
   getParentRoute: () => objectStorageRoute,
   path: 'buckets/create',
 }).lazy(() =>
-  import('./objectStorageLazyRoutes').then(
+  import('src/features/ObjectStorage/objectStorageLandingLazyRoute').then(
     (m) => m.objectStorageLandingLazyRoute
   )
 );
@@ -56,7 +56,7 @@ const objectStorageAccessKeyCreateRoute = createRoute({
   getParentRoute: () => objectStorageRoute,
   path: 'access-keys/create',
 }).lazy(() =>
-  import('./objectStorageLazyRoutes').then(
+  import('src/features/ObjectStorage/objectStorageLandingLazyRoute').then(
     (m) => m.objectStorageLandingLazyRoute
   )
 );
@@ -66,36 +66,36 @@ const objectStorageBucketDetailRoute = createRoute({
   path: 'buckets/$clusterId/$bucketName',
   validateSearch: (search: ObjectStorageDetailSearchParams) => search,
 }).lazy(() =>
-  import('./objectStorageLazyRoutes').then(
-    (m) => m.bucketDetailLandingLazyRoute
-  )
+  import(
+    'src/features/ObjectStorage/BucketDetail/bucketDetailLandingLazyRoute'
+  ).then((m) => m.bucketDetailLandingLazyRoute)
 );
 
 const objectStorageBucketDetailObjectsRoute = createRoute({
   getParentRoute: () => objectStorageBucketDetailRoute,
   path: 'objects',
 }).lazy(() =>
-  import('./objectStorageLazyRoutes').then(
-    (m) => m.bucketDetailLandingLazyRoute
-  )
+  import(
+    'src/features/ObjectStorage/BucketDetail/bucketDetailLandingLazyRoute'
+  ).then((m) => m.bucketDetailLandingLazyRoute)
 );
 
 const objectStorageBucketDetailAccessRoute = createRoute({
   getParentRoute: () => objectStorageBucketDetailRoute,
   path: 'access',
 }).lazy(() =>
-  import('./objectStorageLazyRoutes').then(
-    (m) => m.bucketDetailLandingLazyRoute
-  )
+  import(
+    'src/features/ObjectStorage/BucketDetail/bucketDetailLandingLazyRoute'
+  ).then((m) => m.bucketDetailLandingLazyRoute)
 );
 
 const objectStorageBucketSSLRoute = createRoute({
   getParentRoute: () => objectStorageBucketDetailRoute,
   path: 'ssl',
 }).lazy(() =>
-  import('./objectStorageLazyRoutes').then(
-    (m) => m.bucketDetailLandingLazyRoute
-  )
+  import(
+    'src/features/ObjectStorage/BucketDetail/bucketDetailLandingLazyRoute'
+  ).then((m) => m.bucketDetailLandingLazyRoute)
 );
 export const objectStorageRouteTree = objectStorageRoute.addChildren([
   objectStorageIndexRoute.addChildren([
